[CCAP-567] - create a new Address Utility and move address logic into it
