+++ src/main/java/org/ilgcc/app/submission/actions/SetEditAddressRedirect.java
@@ -3,6 +3,7 @@ package org.ilgcc.app.submission.actions;
 import formflow.library.config.submission.Action;
 import formflow.library.data.Submission;
 import java.util.Map;
+import org.ilgcc.app.utils.AddressUtilities;
 import org.ilgcc.app.utils.SubmissionUtilities;
 import org.springframework.stereotype.Component;
 
@@ -13,7 +14,7 @@ public class SetEditAddressRedirect implements Action {
   public void run(Submission submission) {
     Map<String, Object> inputData = submission.getInputData();
 
-    if (SubmissionUtilities.parentMailingAddressIsHomeAddress(inputData)){
+    if (AddressUtilities.parentMailingAddressIsHomeAddress(inputData)){
       inputData.put("parentEditAddressRedirect", "parent-home-address");
     } else
       inputData.put("parentEditAddressRedirect", "parent-mailing-address");
