+++ src/main/java/org/ilgcc/app/utils/SubmissionUtilities.java
@@ -4,7 +4,6 @@ import static java.lang.Integer.parseInt;
 import static java.util.Collections.emptyList;
 
 import formflow.library.data.Submission;
-import formflow.library.inputs.FieldNameMarkers;
 import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
@@ -50,57 +49,6 @@ public class SubmissionUtilities {
         return submission.getInputData().get(PROGRAM_SCHEDULE).toString();
     }
 
-    /**
-     * @param inputData a JSON object of user inputs
-     * @return true or false
-     */
-    public static boolean parentIsExperiencingHomelessness(Map<String, Object> inputData) {
-        return inputData.getOrDefault("parentHomeExperiencingHomelessness[]", "no").equals(List.of("yes"));
-    }
-
-    /**
-     * @param inputData a JSON object of user inputs
-     * @return true or false
-     */
-    public static boolean parentMailingAddressIsHomeAddress(Map<String, Object> inputData) {
-        return inputData.getOrDefault("parentMailingAddressSameAsHomeAddress[]", "no").equals(List.of("yes"));
-    }
-
-    public static boolean hasAddressSuggestion(Submission submission, String inputName) {
-        return submission.getInputData().get(FieldNameMarkers.UNVALIDATED_FIELD_MARKER_VALIDATE_ADDRESS + inputName)
-                .equals("true") && submission.getInputData()
-                .containsKey(inputName + "StreetAddress1" + FieldNameMarkers.UNVALIDATED_FIELD_MARKER_VALIDATED);
-    }
-
-    public static Map<String, String> getAddress(Map<String, Object> inputData, String addressPrefix) {
-        Map<String, String> addressLines = new HashMap<>();
-
-        String suggestedAddressKey = String.format("useSuggested%sAddress", capitalize(addressPrefix));
-
-        var useSmartyValidatedAddress = inputData.getOrDefault(suggestedAddressKey, "false").equals("true");
-
-        String addressStreet1Key = addressPrefix + (useSmartyValidatedAddress ? "StreetAddress1_validated"
-                : "StreetAddress1");
-        String addressStreet2Key = useSmartyValidatedAddress ? "" : addressPrefix + "StreetAddress2";
-        String cityKey = addressPrefix + (useSmartyValidatedAddress ? "City_validated"
-                : "City");
-        String stateKey = addressPrefix + (useSmartyValidatedAddress ? "State_validated"
-                : "State");
-        String zipCodeKey = addressPrefix + (useSmartyValidatedAddress ? "ZipCode_validated"
-                : "ZipCode");
-
-        addressLines.put("address1", inputData.getOrDefault(addressStreet1Key, "").toString());
-        addressLines.put("address2", inputData.getOrDefault(addressStreet2Key, "").toString());
-        addressLines.put("city", inputData.getOrDefault(cityKey, "").toString());
-        addressLines.put("state", inputData.getOrDefault(stateKey, "").toString());
-        addressLines.put("zipCode", inputData.getOrDefault(zipCodeKey, "").toString());
-
-        return addressLines;
-    }
-
-    private static String capitalize(String text) {
-        return text.substring(0, 1).toUpperCase() + text.substring(1);
-    }
 
     public static String applicantFirstName(Map<String, Object> inputData) {
         var applicantPreferredName = inputData.getOrDefault("parentPreferredName", "").toString();
