+++ src/main/resources/templates/gcc/parent-confirm-address.html
@@ -9,7 +9,7 @@
       <div th:replace="~{fragments/goBack :: goBackLink}"></div>
       <main id="content" role="main" class="form-card spacing-above-35">
         <th:block th:replace="~{fragments/form :: form(action=${formAction}, content=~{::formContent})}">
-          <th:block th:ref="formContent" th:with="hasSuggestion=${T(org.ilgcc.app.utils.SubmissionUtilities).hasAddressSuggestion(submission, 'parentMailing')}">
+          <th:block th:ref="formContent" th:with="hasSuggestion=${T(org.ilgcc.app.utils.AddressUtilities).hasAddressSuggestion(submission, 'parentMailing')}">
             <div class="form-card__content">
               <th:block th:if="${hasSuggestion}">
                 <th:block th:replace="~{fragments/inputs/overrides/addressSuggestionFound ::
