+++ src/main/resources/templates/providerresponse/confirm-address.html
@@ -11,7 +11,7 @@
                 <th:block
                         th:replace="~{fragments/form :: form(action=${formAction}, content=~{::formContent})}">
                     <th:block th:ref="formContent"
-                              th:with="hasSuggestion=${T(org.ilgcc.app.utils.SubmissionUtilities).hasAddressSuggestion(submission, 'providerResponseService')}">
+                              th:with="hasSuggestion=${T(org.ilgcc.app.utils.AddressUtilities).hasAddressSuggestion(submission, 'providerResponseService')}">
                         <div class="form-card__content">
                             <th:block th:if="${hasSuggestion}">
                                 <th:block th:replace="~{fragments/inputs/overrides/addressSuggestionFound ::
