+++ src/main/resources/templates/providerresponse/registration-confirm-mailing-address.html
@@ -11,7 +11,7 @@
                 <th:block
                         th:replace="~{fragments/form :: form(action=${formAction}, content=~{::formContent})}">
                     <th:block th:ref="formContent"
-                              th:with="hasSuggestion=${T(org.ilgcc.app.utils.SubmissionUtilities).hasAddressSuggestion(submission, 'providerMailing')}">
+                              th:with="hasSuggestion=${T(org.ilgcc.app.utils.AddressUtilities).hasAddressSuggestion(submission, 'providerMailing')}">
                         <div class="form-card__content">
                             <th:block th:if="${hasSuggestion}">
                                 <th:block th:replace="~{fragments/inputs/overrides/addressSuggestionFound ::
